requirements.txt
fitz==0.0.1.dev2
streamlit>=1.20.0
google-generativeai>=0.1.0
PyMuPDF>=1.22.0
PyPDF2
transformers

